refactor: import compat helpers from abicheck.compat.cli in tests
Tests were importing private compat helpers via re-export from
abicheck.cli. Since compat is a separate module, tests should import
directly from abicheck.compat.cli. Remove the re-export from cli.py.
