+++ tests/test_abicc_parity_extended.py
@@ -25,11 +25,8 @@ from pathlib import Path
 import pytest
 
 from abicheck.checker import Change, ChangeKind, DiffResult, Verdict
-from abicheck.cli import (
-    _apply_strict,
-    _filter_source_only,
-    main,
-)
+from abicheck.cli import main
+from abicheck.compat.cli import _apply_strict, _filter_source_only
 
 # ---------------------------------------------------------------------------
 # Shared helpers
